[Bronze III] Title: 주차 요금 정산하기, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/33753. 주차 요금 정산하기/주차 요금 정산하기.cc
@@ -8,5 +8,5 @@ int main() {
 	int t;
 	cin >> t;
 	if (t <= 30)cout << a;
-	else cout << a + (ceil(double(t-30) / b)) * c;
+	else cout << a + int(ceil(double(t-30) / b)) * c;
 }
